fix(cloud): Switch positive values to unsigned
This adds the benefit of catching errors early

+++ internal/cli/kraft/cloud/compose/up/up.go
@@ -226,7 +226,7 @@ func Up(ctx context.Context, opts *UpOptions, args ...string) error {
 			Client:                 opts.Client,
 			Env:                    env,
 			Image:                  service.Image,
-			Memory:                 memory,
+			Memory:                 uint(memory),
 			Metro:                  opts.Metro,
 			Name:                   service.Name,
 			ServiceGroupNameOrUUID: serviceGroup,
